+++ tests/test_github.py
@@ -357,4 +357,4 @@ class TestGitHub(BaseTest):
         assert_failure(["github", "update-pr-descriptions", "--all", "--related"],
                        "update-pr-descriptions subcommand must take exactly one of the following options: --all, --mine, --related")
         assert_failure(["github", "update-pr-descriptions", "--update-related-descriptions"],
-                       "--update-related-descriptions option is only valid with create-pr subcommand.")
+                       "--update-related-descriptions option is only valid with create-pr, restack-pr and retarget-pr subcommands.")