+++ tests/test_github_retarget_pr.py
@@ -424,10 +424,16 @@ class TestGitHubRetargetPR(BaseTest):
         self.repo_sandbox.set_git_config_key("machete.github.prDescriptionIntroStyle", "full")
 
         assert_success(
-            ['github', 'retarget-pr'],
-            'Base branch of PR #15 has been switched to branch-1\n'
-            'Checking for open GitHub PRs... OK\n'
-            'Description of PR #15 has been updated\n'
+            ['github', 'retarget-pr', '-U'],
+            """
+            Base branch of PR #15 has been switched to branch-1
+            Checking for open GitHub PRs... OK
+            Description of PR #15 has been updated
+            Updating descriptions of other PRs...
+            Description of PR #20 (feature_1 -> feature) has been updated
+            Description of PR #25 (feature_2 -> feature) has been updated
+            Description of PR #35 (feature_4 -> feature) has been updated
+            """
         )
         pr15 = github_api_state.get_pull_by_number(15)
         assert pr15 is not None